requirements.txt
Django==1.11.4
django-crispy-forms==1.7.1
djangorestframework==3.7.7
pytz==2018.3
recaptcha-client==1.0.6
